requirements.txt
newspaper3k
flask
